requirements.txt
adrf==0.1.12
asgiref==3.11.0
async-property==0.2.2
click==8.3.1
colorama==0.4.6
Django==6.0
djangorestframework==3.16.1
h11==0.16.0
pillow==12.0.0
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
